refactor and split tasks
